remove windows target; clean-up
