+++ src/utils.rs
@@ -1,7 +1,9 @@
 use std::ffi::OsStr;
-use std::io::Read;
+use std::io::{IoSliceMut, Read};
 use std::path::Path;
 use std::str::FromStr;
+
+use indicatif::{ProgressBar, ProgressBarIter, ProgressStyle};
 
 use crate::append_vec::{AppendVec, StoredAccountMeta};
 
@@ -59,4 +61,56 @@ pub(crate) trait ReadProgressTracking {
         rd: Box<dyn Read>,
         file_len: u64,
     ) -> Box<dyn Read>;
+}
+
+pub(crate) struct LoadProgressTracking {}
+
+impl ReadProgressTracking for LoadProgressTracking {
+    fn new_read_progress_tracker(
+        &self,
+        _path: &Path,
+        rd: Box<dyn Read>,
+        file_len: u64,
+    ) -> Box<dyn Read> {
+        let progress_bar = ProgressBar::new(file_len).with_style(
+            ProgressStyle::with_template(
+                "{prefix:>15.bold.dim} {spinner:.green} [{bar:.cyan/blue}] {bytes}/{total_bytes} \
+                 ({percent}%)",
+            )
+            .unwrap()
+            .progress_chars("#>-"),
+        );
+        progress_bar.set_prefix("manifest");
+
+        Box::new(LoadProgressTracker { rd: progress_bar.wrap_read(rd), progress_bar })
+    }
+}
+
+struct LoadProgressTracker {
+    progress_bar: ProgressBar,
+    rd: ProgressBarIter<Box<dyn Read>>,
+}
+
+impl Drop for LoadProgressTracker {
+    fn drop(&mut self) {
+        self.progress_bar.finish()
+    }
+}
+
+impl Read for LoadProgressTracker {
+    fn read(&mut self, buf: &mut [u8]) -> std::io::Result<usize> {
+        self.rd.read(buf)
+    }
+
+    fn read_vectored(&mut self, bufs: &mut [IoSliceMut<'_>]) -> std::io::Result<usize> {
+        self.rd.read_vectored(bufs)
+    }
+
+    fn read_to_string(&mut self, buf: &mut String) -> std::io::Result<usize> {
+        self.rd.read_to_string(buf)
+    }
+
+    fn read_exact(&mut self, buf: &mut [u8]) -> std::io::Result<()> {
+        self.rd.read_exact(buf)
+    }
 }
